chore: improve code layout
